+++ api-types/src/lib.rs
@@ -9,6 +9,7 @@ mod vm_resize;
 mod vm_resize_disk;
 mod vm_resize_zone;
 mod vm_snapshot;
+mod vmm_ping_response;
 
 pub use vm_coredump::VmCoredumpData;
 pub use vm_remove_device::VmRemoveDeviceData;
@@ -16,3 +17,4 @@ pub use vm_resize::VmResizeData;
 pub use vm_resize_disk::VmResizeDiskData;
 pub use vm_resize_zone::VmResizeZoneData;
 pub use vm_snapshot::VmSnapshotConfig;
+pub use vmm_ping_response::VmmPingResponse;
